leona-test-folder/studio/requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic
typing-extensions

